+++ data_processing/radiology/refined_table/annotation/generate.py
@@ -18,7 +18,7 @@ from data_processing.common.config import ConfigSet
 from data_processing.common.sparksession import SparkConfig
 from data_processing.common.custom_logger import init_logger
 import data_processing.common.constants as const
-from data_processing.radiology.common.utils import overlay_images, create_seg_images
+from data_processing.radiology.common.preprocess import overlay_images, create_seg_images
 
 from pyspark.sql import functions as F
 from pyspark.sql.types import StringType, IntegerType, ArrayType, StructType, StructField, BinaryType
@@ -88,6 +88,8 @@ def generate_png_tables(cfg):
         seg_png_table_path = const.TABLE_LOCATION(cfg)
         
         # find images with tumor
+        spark.sparkContext.addPyFile("./data_processing/radiology/common/preprocess.py")
+        from preprocess import create_seg_images, overlay_images
         create_seg_png_udf = F.udf(create_seg_images, ArrayType(StructType(
                                     [StructField("instance_number", IntegerType()),
                                      StructField("scan_annotation_record_uuid", StringType()),
@@ -124,6 +126,8 @@ def generate_png_tables(cfg):
 
         # generate uuid
         spark.sparkContext.addPyFile("./data_processing/common/EnsureByteContext.py")
+        spark.sparkContext.addPyFile("./data_processing/common/utils.py")
+        from utils import generate_uuid_binary
         generate_uuid_udf = F.udf(generate_uuid_binary, StringType())
         seg_df = seg_df.withColumn("png_record_uuid", F.lit(generate_uuid_udf(seg_df.overlay, F.array([F.lit("PNG")]))))
